+++ CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - start: overwrite /etc/hosts of a pot, adding all pots on the same bridge and custom entries added via set-hosts
 - flavorable commands: extend support to set-cmd and set-env
 - pot-rdr anchor: the name of the anchor is now a truncated pot name (the last 54 characters)
-- export: it's executed only if one snapshot is available. -F force execution, -A call purge-snapshots to reduce the number of shapshot to 1
+- export: it's executed only if one snapshot is available. -F force execution, -A fix the number of snapshots, via purge-snapshots or taking a shapshot automagically
 - start: using exec.start instead of command (it seems more predictable)
 
 ### Deprecated